Created meta-client to take care of OAuth
